feat(fullscreenVideoCall): responsive video controls
