+++ src/script/components/calling/FullscreenVideoCall.tsx
@@ -17,24 +17,14 @@
  *
  */
 
-import React, {useEffect, useRef, useState} from 'react';
+import React, {useEffect, useState} from 'react';
 
 import {DefaultConversationRoleName} from '@wireapp/api-client/lib/conversation/';
 import {TabIndex} from '@wireapp/react-ui-kit/lib/types/enums';
 import cx from 'classnames';
 import {container} from 'tsyringe';
 
-import {CALL_TYPE} from '@wireapp/avs';
-import {
-  Checkbox,
-  CheckboxLabel,
-  EmojiIcon,
-  GridIcon,
-  IconButton,
-  IconButtonVariant,
-  RaiseHandIcon,
-  Select,
-} from '@wireapp/react-ui-kit';
+import {Checkbox, CheckboxLabel, IconButton, IconButtonVariant} from '@wireapp/react-ui-kit';
 import {WebAppEvents} from '@wireapp/webapp-events';
 
 import {useAppNotification} from 'Components/AppNotification/AppNotification';
@@ -42,17 +32,13 @@ import {useCallAlertState} from 'Components/calling/useCallAlertState';
 import {ConversationClassifiedBar} from 'Components/ClassifiedBar/ClassifiedBar';
 import * as Icon from 'Components/Icon';
 import {ModalComponent} from 'Components/Modals/ModalComponent';
-import {useClickOutside} from 'Hooks/useClickOutside';
 import {CallingRepository} from 'src/script/calling/CallingRepository';
-import {Config} from 'src/script/Config';
-import {isCallViewOption} from 'src/script/guards/CallView';
-import {isMediaDevice} from 'src/script/guards/MediaDevice';
 import {useActiveWindowMatchMedia} from 'src/script/hooks/useActiveWindowMatchMedia';
 import {useToggleState} from 'src/script/hooks/useToggleState';
-import {MediaDeviceType} from 'src/script/media/MediaDeviceType';
+import {CallViewTab} from 'src/script/view_model/CallingViewModel';
 import {useKoSubscribableChildren} from 'Util/ComponentUtil';
 import {isDetachedCallingFeatureEnabled} from 'Util/isDetachedCallingFeatureEnabled';
-import {handleKeyDown, isEscapeKey} from 'Util/KeyboardUtil';
+import {handleKeyDown} from 'Util/KeyboardUtil';
 import {t} from 'Util/LocalizerUtil';
 import {preventFocusOutside} from 'Util/util';
 
@@ -63,30 +49,23 @@ import {
   headerActionsWrapperStyles,
   paginationButtonStyles,
   paginationWrapperStyles,
-  videoControlActiveStyles,
-  videoControlDisabledStyles,
   videoControlInActiveStyles,
   videoTopBarStyles,
 } from './FullscreenVideoCall.styles';
 import {GroupVideoGrid} from './GroupVideoGrid';
 import {Pagination} from './Pagination';
 import {useSyncCurrentRange} from './useSyncCurrentRange';
+import {VideoControls} from './VideoControls/VideoControls';
 
 import type {Call} from '../../calling/Call';
 import {CallingViewMode, CallState, MuteState} from '../../calling/CallState';
 import {Participant} from '../../calling/Participant';
 import type {Grid} from '../../calling/videoGridHandler';
 import type {Conversation} from '../../entity/Conversation';
-import {ElectronDesktopCapturerSource, MediaDevicesHandler} from '../../media/MediaDevicesHandler';
+import {MediaDevicesHandler} from '../../media/MediaDevicesHandler';
 import {TeamState} from '../../team/TeamState';
-import {CallViewTab} from '../../view_model/CallingViewModel';
 import {useWarningsState} from '../../view_model/WarningsContainer/WarningsState';
 import {CONFIG, TYPE} from '../../view_model/WarningsContainer/WarningsTypes';
-
-enum BlurredBackgroundStatus {
-  OFF = 'bluroff',
-  ON = 'bluron',
-}
 
 export interface FullscreenVideoCallProps {
   activeCallViewTab: string;
@@ -116,8 +95,6 @@ export interface FullscreenVideoCallProps {
   sendHandRaised: (isHandUp: boolean, call: Call) => void;
   videoGrid: Grid;
 }
-
-const EMOJIS_LIST = ['👍', '🎉', '❤️', '😂', '😮', '👏', '🤔', '😢', '👎'];
 
 const LOCAL_STORAGE_KEY_FOR_SCREEN_SHARING_CONFIRM_MODAL = 'DO_NOT_ASK_AGAIN_FOR_SCREEN_SHARING_CONFIRM_MODAL';
 
@@ -152,17 +129,8 @@ const FullscreenVideoCall = ({
   callState = container.resolve(CallState),
 }: FullscreenVideoCallProps) => {
   const [isConfirmCloseModalOpen, setIsConfirmCloseModalOpen] = useState<boolean>(false);
-  const [showEmojisBar, setShowEmojisBar] = useState<boolean>(false);
-  const [disabledEmojis, setDisabledEmojis] = useState<string[]>([]);
   const selfParticipant = call.getSelfParticipant();
-  const {
-    sharesScreen: selfSharesScreen,
-    sharesCamera: selfSharesCamera,
-    handRaisedAt: selfHandRaisedAt,
-  } = useKoSubscribableChildren(selfParticipant, ['sharesScreen', 'sharesCamera', 'handRaisedAt']);
-  const isSelfHandRaised = Boolean(selfHandRaisedAt);
-  const emojiBarRef = useRef(null);
-  const emojiBarToggleButtonRef = useRef(null);
+  const {sharesCamera: selfSharesCamera} = useKoSubscribableChildren(selfParticipant, ['sharesScreen', 'sharesCamera']);
 
   // Warnings banner
   const warnings = useWarningsState(state => state.warnings);
@@ -170,9 +138,6 @@ const FullscreenVideoCall = ({
   const isConnectivityRecovery = visibleWarning === TYPE.CONNECTIVITY_RECOVERY;
   const hasOffset = warnings.length > 0 && !isConnectivityRecovery;
   const isMiniMode = CONFIG.MINI_MODES.includes(visibleWarning);
-
-  const {blurredVideoStream} = useKoSubscribableChildren(selfParticipant, ['blurredVideoStream']);
-  const hasBlurredBackground = !!blurredVideoStream;
 
   const {
     activeSpeakers,
@@ -189,41 +154,14 @@ const FullscreenVideoCall = ({
     'participants',
     'handRaisedParticipants',
   ]);
-  const {display_name: conversationName, is1to1: is1to1Conversation} = useKoSubscribableChildren(conversation, [
-    'display_name',
-    'is1to1',
-  ]);
-  const {isVideoCallingEnabled, classifiedDomains} = useKoSubscribableChildren(teamState, [
-    'isVideoCallingEnabled',
-    'classifiedDomains',
-  ]);
-
-  const {
-    [MediaDeviceType.VIDEO_INPUT]: currentCameraDevice,
-    [MediaDeviceType.AUDIO_INPUT]: currentMicrophoneDevice,
-    [MediaDeviceType.AUDIO_OUTPUT]: currentSpeakerDevice,
-  } = useKoSubscribableChildren(mediaDevicesHandler.currentDeviceId, [
-    MediaDeviceType.VIDEO_INPUT,
-    MediaDeviceType.AUDIO_INPUT,
-    MediaDeviceType.AUDIO_OUTPUT,
-  ]);
-
-  const {videoinput, audioinput, audiooutput} = useKoSubscribableChildren(mediaDevicesHandler.availableDevices, [
-    MediaDeviceType.VIDEO_INPUT,
-    MediaDeviceType.AUDIO_INPUT,
-    MediaDeviceType.AUDIO_OUTPUT,
-  ]);
+  const {display_name: conversationName} = useKoSubscribableChildren(conversation, ['display_name']);
+  const {classifiedDomains} = useKoSubscribableChildren(teamState, ['classifiedDomains']);
 
   const {selfUser, roles} = useKoSubscribableChildren(conversation, ['selfUser', 'roles']);
   const {emojis, viewMode, detachedWindow, isScreenSharingSourceFromDetachedWindow} = useKoSubscribableChildren(
     callState,
     ['emojis', 'viewMode', 'detachedWindow', 'isScreenSharingSourceFromDetachedWindow'],
   );
-
-  useClickOutside(emojiBarRef, () => setShowEmojisBar(false), emojiBarToggleButtonRef, detachedWindow?.document);
-
-  const [audioOptionsOpen, setAudioOptionsOpen] = useState(false);
-  const [videoOptionsOpen, setVideoOptionsOpen] = useState(false);
 
   const minimize = () => {
     const isSharingScreen = call?.getSelfParticipant().sharesScreen();
@@ -239,7 +177,6 @@ const FullscreenVideoCall = ({
   };
   const openPopup = () => callingRepository.setViewModeDetached();
 
-  const [isCallViewOpen, toggleCallView] = useToggleState(false);
   const [isParticipantsListOpen, toggleParticipantsList] = useToggleState(false);
 
   const handRaisedNotification = useAppNotification({
@@ -284,165 +221,12 @@ const FullscreenVideoCall = ({
     };
   }, [viewMode]);
 
-  const showToggleVideo =
-    isVideoCallingEnabled &&
-    (call.initialType === CALL_TYPE.VIDEO || conversation.supportsVideoCall(call.isConference));
-
-  const showSwitchMicrophone = audioinput.length > 1;
-
-  const callViewOptions = [
-    {
-      label: t('videoCallOverlayViewModeLabel'),
-      options: [
-        {
-          label: t('videoCallOverlayViewModeAll'),
-          value: CallViewTab.ALL,
-        },
-        {
-          label: t('videoCallOverlayViewModeSpeakers'),
-          value: CallViewTab.SPEAKERS,
-        },
-      ],
-    },
-  ];
-
-  const selectedCallViewOption =
-    callViewOptions[0].options.find(option => option.value === activeCallViewTab) ?? callViewOptions[0].options[0];
-
-  const audioOptions = [
-    {
-      label: t('videoCallaudioInputMicrophone'),
-      options: audioinput.map((device: MediaDeviceInfo | ElectronDesktopCapturerSource) => {
-        return isMediaDevice(device)
-          ? {
-              label: device.label,
-              value: `${device.deviceId}-input`,
-              dataUieName: `${device.deviceId}-input`,
-              id: device.deviceId,
-            }
-          : {
-              label: device.name,
-              value: `${device.id}-input`,
-              dataUieName: `${device.id}-input`,
-              id: device.id,
-            };
-      }),
-    },
-    {
-      label: t('videoCallaudioOutputSpeaker'),
-      options: audiooutput.map((device: MediaDeviceInfo | ElectronDesktopCapturerSource) => {
-        return isMediaDevice(device)
-          ? {
-              label: device.label,
-              value: `${device.deviceId}-output`,
-              dataUieName: `${device.deviceId}-output`,
-              id: device.deviceId,
-            }
-          : {
-              label: device.name,
-              value: `${device.id}-output`,
-              dataUieName: `${device.id}-output`,
-              id: device.id,
-            };
-      }),
-    },
-  ];
-
-  const [selectedAudioOptions, setSelectedAudioOptions] = useState(() =>
-    [currentMicrophoneDevice, currentSpeakerDevice].flatMap(
-      (device, index) => audioOptions[index].options.find(item => item.id === device) ?? audioOptions[index].options[0],
-    ),
-  );
-
-  const updateAudioOptions = (selectedOption: string, input: boolean) => {
-    const microphone = input
-      ? audioOptions[0].options.find(item => item.value === selectedOption) ?? selectedAudioOptions[0]
-      : selectedAudioOptions[0];
-    const speaker = !input
-      ? audioOptions[1].options.find(item => item.value === selectedOption) ?? selectedAudioOptions[1]
-      : selectedAudioOptions[1];
-
-    setSelectedAudioOptions([microphone, speaker]);
-    switchMicrophoneInput(microphone.id);
-    switchSpeakerOutput(speaker.id);
-  };
-
-  const isBlurredBackgroundEnabled = Config.getConfig().FEATURE.ENABLE_BLUR_BACKGROUND;
-
-  const blurredBackgroundOptions = {
-    label: t('videoCallbackgroundBlurHeadline'),
-    options: [
-      {
-        // Blurring is not possible if webgl context is not available
-        isDisabled: !document.createElement('canvas').getContext('webgl2'),
-        label: t('videoCallbackgroundBlur'),
-        value: BlurredBackgroundStatus.ON,
-        dataUieName: 'blur',
-        id: BlurredBackgroundStatus.ON,
-      },
-      {
-        label: t('videoCallbackgroundNotBlurred'),
-        value: BlurredBackgroundStatus.OFF,
-        dataUieName: 'no-blur',
-        id: BlurredBackgroundStatus.OFF,
-      },
-    ],
-  };
-
-  const videoOptions = [
-    {
-      label: t('videoCallvideoInputCamera'),
-      options: videoinput.map((device: MediaDeviceInfo | ElectronDesktopCapturerSource) => {
-        return isMediaDevice(device)
-          ? {
-              label: device.label,
-              value: device.deviceId,
-              dataUieName: device.deviceId,
-              id: device.deviceId,
-            }
-          : {
-              label: device.name,
-              value: device.id,
-              dataUieName: device.id,
-              id: device.id,
-            };
-      }),
-    },
-    ...(isBlurredBackgroundEnabled ? [blurredBackgroundOptions] : []),
-  ];
-
-  const selectedVideoOptions = [currentCameraDevice, hasBlurredBackground]
-    .flatMap(device => videoOptions.flatMap(options => options.options.filter(item => item.id === device)) ?? [])
-    .concat(hasBlurredBackground ? blurredBackgroundOptions.options[0] : blurredBackgroundOptions.options[1]);
-
-  const updateVideoOptions = (selectedOption: string | BlurredBackgroundStatus) => {
-    const camera = videoOptions[0].options.find(item => item.value === selectedOption) ?? selectedVideoOptions[0];
-    if (selectedOption === BlurredBackgroundStatus.ON) {
-      switchBlurredBackground(true);
-    } else if (selectedOption === BlurredBackgroundStatus.OFF) {
-      switchBlurredBackground(false);
-    } else {
-      switchCameraInput(camera.id);
-    }
-  };
-
-  const onEmojiClick = (selectedEmoji: string) => {
-    setDisabledEmojis(prev => [...prev, selectedEmoji]);
-
-    sendEmoji(selectedEmoji, call);
-
-    setTimeout(() => {
-      setDisabledEmojis(prev => [...prev].filter(emoji => emoji !== selectedEmoji));
-    }, CallingRepository.EMOJI_TIME_OUT_DURATION);
-  };
-
   const {showAlert, isGroupCall, clearShowAlert} = useCallAlertState();
 
   const totalPages = callPages.length;
 
   // To be changed when design chooses a breakpoint, the conditional can be integrated to the ui-kit directly
   const horizontalSmBreakpoint = useActiveWindowMatchMedia('max-width: 680px');
-  const horizontalXsBreakpoint = useActiveWindowMatchMedia('max-width: 500px');
 
   const callGroupStartedAlert = t(isGroupCall ? 'startedVideoGroupCallingAlert' : 'startedVideoCallingAlert', {
     conversationName,
@@ -653,332 +437,29 @@ const FullscreenVideoCall = ({
                 </span>
               </div>
             ))}
-
-            <ul className="video-controls__wrapper">
-              {!horizontalSmBreakpoint && (
-                <li className="video-controls__item__minimize">
-                  <button
-                    className="video-controls__button video-controls__button--small"
-                    css={videoControlInActiveStyles}
-                    onClick={minimize}
-                    onKeyDown={event => handleKeyDown(event, () => minimize())}
-                    type="button"
-                    data-uie-name="do-call-controls-video-minimize"
-                    title={t('videoCallOverlayCloseFullScreen')}
-                  >
-                    {viewMode === CallingViewMode.DETACHED_WINDOW ? (
-                      <Icon.CloseDetachedWindowIcon />
-                    ) : (
-                      <Icon.MessageIcon />
-                    )}
-                  </button>
-                </li>
-              )}
-              <div className="video-controls__centered-items">
-                <li className="video-controls__item">
-                  <button
-                    className="video-controls__button"
-                    data-uie-value={!isMuted ? 'inactive' : 'active'}
-                    onClick={() => toggleMute(call, !isMuted)}
-                    onKeyDown={event => handleKeyDown(event, () => toggleMute(call, !isMuted))}
-                    css={!isMuted ? videoControlActiveStyles : videoControlInActiveStyles}
-                    type="button"
-                    data-uie-name="do-call-controls-video-call-mute"
-                    role="switch"
-                    aria-checked={!isMuted}
-                    title={t('videoCallOverlayMicrophone')}
-                  >
-                    {isMuted ? <Icon.MicOffIcon width={16} height={16} /> : <Icon.MicOnIcon width={16} height={16} />}
-                  </button>
-
-                  {showSwitchMicrophone && (
-                    <button
-                      className="device-toggle-button"
-                      css={audioOptionsOpen ? videoControlActiveStyles : videoControlInActiveStyles}
-                      onClick={() => setAudioOptionsOpen(prev => !prev)}
-                      onKeyDown={event => handleKeyDown(event, () => setAudioOptionsOpen(prev => !prev))}
-                      onBlur={event => {
-                        if (!event.currentTarget.contains(event.relatedTarget)) {
-                          setAudioOptionsOpen(false);
-                        }
-                      }}
-                    >
-                      {audioOptionsOpen ? (
-                        <>
-                          <Select
-                            // eslint-disable-next-line jsx-a11y/no-autofocus
-                            autoFocus
-                            value={selectedAudioOptions}
-                            id="select-microphone"
-                            dataUieName="select-microphone"
-                            controlShouldRenderValue={false}
-                            isClearable={false}
-                            backspaceRemovesValue={false}
-                            hideSelectedOptions={false}
-                            options={audioOptions}
-                            onChange={selectedOption => {
-                              updateAudioOptions(
-                                String(selectedOption?.value),
-                                String(selectedOption?.value).includes('input'),
-                              );
-                            }}
-                            onKeyDown={event => isEscapeKey(event) && setAudioOptionsOpen(false)}
-                            menuPlacement="top"
-                            menuIsOpen
-                            wrapperCSS={{marginBottom: 0}}
-                          />
-                          <Icon.ChevronIcon css={{height: '16px'}} />
-                        </>
-                      ) : (
-                        <Icon.ChevronIcon css={{rotate: '180deg', height: '16px'}} />
-                      )}
-                    </button>
-                  )}
-                </li>
-
-                {showToggleVideo && (
-                  <li className="video-controls__item">
-                    <button
-                      className="video-controls__button"
-                      data-uie-value={selfSharesCamera ? 'active' : 'inactive'}
-                      onClick={() => toggleCamera(call)}
-                      onKeyDown={event => handleKeyDown(event, () => toggleCamera(call))}
-                      role="switch"
-                      aria-checked={selfSharesCamera}
-                      tabIndex={TabIndex.FOCUSABLE}
-                      css={selfSharesCamera ? videoControlActiveStyles : videoControlInActiveStyles}
-                      data-uie-name="do-call-controls-toggle-video"
-                      title={t('videoCallOverlayCamera')}
-                    >
-                      {selfSharesCamera ? (
-                        <Icon.CameraIcon width={16} height={16} />
-                      ) : (
-                        <Icon.CameraOffIcon width={16} height={16} />
-                      )}
-                    </button>
-
-                    <button
-                      className="device-toggle-button"
-                      css={videoOptionsOpen ? videoControlActiveStyles : videoControlInActiveStyles}
-                      onClick={() => setVideoOptionsOpen(prev => !prev)}
-                      onKeyDown={event => handleKeyDown(event, () => setVideoOptionsOpen(prev => !prev))}
-                      onBlur={event => {
-                        if (!event.currentTarget.contains(event.relatedTarget)) {
-                          setVideoOptionsOpen(false);
-                        }
-                      }}
-                    >
-                      {videoOptionsOpen ? (
-                        <>
-                          <Select
-                            // eslint-disable-next-line jsx-a11y/no-autofocus
-                            autoFocus
-                            value={selectedVideoOptions}
-                            onChange={selectedOption => updateVideoOptions(String(selectedOption?.value))}
-                            onKeyDown={event => isEscapeKey(event) && setVideoOptionsOpen(false)}
-                            id="select-camera"
-                            dataUieName="select-camera"
-                            controlShouldRenderValue={false}
-                            isClearable={false}
-                            backspaceRemovesValue={false}
-                            hideSelectedOptions={false}
-                            options={videoOptions}
-                            menuPlacement="top"
-                            menuIsOpen
-                            wrapperCSS={{marginBottom: 0}}
-                          />
-                          <Icon.ChevronIcon css={{height: '16px'}} />
-                        </>
-                      ) : (
-                        <Icon.ChevronIcon css={{rotate: '180deg', height: '16px'}} />
-                      )}
-                    </button>
-                  </li>
-                )}
-
-                <li className="video-controls__item">
-                  <button
-                    className={`video-controls__button ${!canShareScreen ? 'with-tooltip with-tooltip--top' : ''}`}
-                    data-tooltip={t('videoCallScreenShareNotSupported')}
-                    css={
-                      !canShareScreen
-                        ? videoControlDisabledStyles
-                        : selfSharesScreen
-                          ? videoControlActiveStyles
-                          : videoControlInActiveStyles
-                    }
-                    onClick={() => toggleScreenshare(call)}
-                    onKeyDown={event => handleKeyDown(event, () => toggleScreenshare(call))}
-                    type="button"
-                    data-uie-value={selfSharesScreen ? 'active' : 'inactive'}
-                    data-uie-enabled={canShareScreen ? 'true' : 'false'}
-                    data-uie-name="do-toggle-screen"
-                    title={t('videoCallOverlayShareScreen')}
-                  >
-                    {selfSharesScreen ? (
-                      <Icon.ScreenshareIcon width={16} height={16} />
-                    ) : (
-                      <Icon.ScreenshareOffIcon width={16} height={16} />
-                    )}
-                  </button>
-                </li>
-
-                <li className="video-controls__item">
-                  <button
-                    className="video-controls__button video-controls__button--red"
-                    onClick={() => leave(call)}
-                    onKeyDown={event => handleKeyDown(event, () => leave(call))}
-                    type="button"
-                    data-uie-name="do-call-controls-video-call-cancel"
-                    title={t('videoCallOverlayHangUp')}
-                  >
-                    <Icon.HangupIcon />
-                  </button>
-                </li>
-              </div>
-
-              {!horizontalXsBreakpoint && (
-                <div
-                  css={{
-                    ...(!horizontalSmBreakpoint && {minWidth: '157px'}),
-                    display: 'flex',
-                    justifyContent: 'flex-end',
-                    gap: 10,
-                    marginRight: 10,
-                  }}
-                >
-                  {participants.length > 2 && (
-                    <li className="video-controls__item">
-                      <button
-                        onBlur={event => {
-                          if (!event.currentTarget.contains(event.relatedTarget) && isCallViewOpen) {
-                            toggleCallView();
-                          }
-                        }}
-                        className={cx('video-controls__button_primary', {active: isCallViewOpen})}
-                        onClick={toggleCallView}
-                        onKeyDown={event => handleKeyDown(event, toggleCallView)}
-                        type="button"
-                        data-uie-name="do-call-controls-video-call-view"
-                        role="switch"
-                        aria-checked={!isMuted}
-                        title={t('videoCallOverlayChangeViewMode')}
-                      >
-                        {isCallViewOpen && (
-                          <Select
-                            // eslint-disable-next-line jsx-a11y/no-autofocus
-                            autoFocus
-                            value={selectedCallViewOption}
-                            id="select-call-view"
-                            dataUieName="select-call-view"
-                            controlShouldRenderValue={false}
-                            isClearable={false}
-                            backspaceRemovesValue={false}
-                            hideSelectedOptions={false}
-                            options={callViewOptions}
-                            onChange={selectedOption => {
-                              if (isCallViewOption(selectedOption)) {
-                                setActiveCallViewTab(selectedOption.value);
-                                setMaximizedParticipant(call, null);
-                              }
-                            }}
-                            onKeyDown={event => isEscapeKey(event) && setAudioOptionsOpen(false)}
-                            menuPlacement="top"
-                            menuIsOpen={isCallViewOpen}
-                            wrapperCSS={{marginBottom: 0, width: 0, height: 0}}
-                            menuCSS={{right: 0, bottom: 10}}
-                          />
-                        )}
-                        <GridIcon width={16} height={16} />
-                      </button>
-                    </li>
-                  )}
-
-                  {Config.getConfig().FEATURE.ENABLE_IN_CALL_HAND_RAISE && !is1to1Conversation && (
-                    <li className="video-controls__item">
-                      <button
-                        data-uie-value={isSelfHandRaised ? 'active' : 'inactive'}
-                        onClick={() => toggleIsHandRaised(isSelfHandRaised)}
-                        onKeyDown={event => handleKeyDown(event, () => toggleIsHandRaised(isSelfHandRaised))}
-                        className={cx('video-controls__button_primary', {active: isSelfHandRaised})}
-                        type="button"
-                        data-uie-name="do-toggle-hand-raise"
-                        role="switch"
-                        aria-checked={isSelfHandRaised}
-                        title={
-                          isSelfHandRaised
-                            ? t('videoCallParticipantLowerYourHand')
-                            : t('videoCallParticipantRaiseYourHand')
-                        }
-                      >
-                        <RaiseHandIcon width={16} height={16} />
-                      </button>
-                    </li>
-                  )}
-
-                  {Config.getConfig().FEATURE.ENABLE_IN_CALL_REACTIONS && (
-                    <li className="video-controls__item">
-                      {showEmojisBar && (
-                        <div
-                          ref={emojiBarRef}
-                          role="toolbar"
-                          className="video-controls-emoji-bar"
-                          data-uie-name="video-controls-emoji-bar"
-                          aria-label={t('callReactionButtonsAriaLabel')}
-                        >
-                          {EMOJIS_LIST.map(emoji => {
-                            const isDisabled = disabledEmojis.includes(emoji);
-                            return (
-                              <button
-                                aria-label={t('callReactionButtonAriaLabel', {emoji})}
-                                data-uie-name="video-controls-emoji"
-                                data-uie-value={emoji}
-                                key={emoji}
-                                disabled={isDisabled}
-                                onClick={() => onEmojiClick(emoji)}
-                                className={cx({disabled: isDisabled})}
-                              >
-                                {emoji}
-                              </button>
-                            );
-                          })}
-                        </div>
-                      )}
-                      <button
-                        ref={emojiBarToggleButtonRef}
-                        title={t('callReactions')}
-                        className={cx('video-controls__button_primary', {active: showEmojisBar})}
-                        onClick={() => setShowEmojisBar(prev => !prev)}
-                        type="button"
-                        aria-labelledby="show-emoji-bar"
-                        data-uie-name="do-call-controls-video-call-cancel"
-                      >
-                        <EmojiIcon />
-                      </button>
-                    </li>
-                  )}
-                  <li className="video-controls__item">
-                    <button
-                      data-uie-value={isParticipantsListOpen ? 'active' : 'inactive'}
-                      onClick={toggleParticipantsList}
-                      onKeyDown={event => handleKeyDown(event, toggleParticipantsList)}
-                      className={cx('video-controls__button_primary', {active: isParticipantsListOpen})}
-                      type="button"
-                      data-uie-name="do-toggle-call-participants-list"
-                      role="switch"
-                      aria-checked={isParticipantsListOpen}
-                      title={
-                        isParticipantsListOpen
-                          ? t('videoCallOverlayHideParticipantsList')
-                          : t('videoCallOverlayShowParticipantsList')
-                      }
-                    >
-                      <Icon.PeopleIcon width={16} height={16} />
-                    </button>
-                  </li>
-                </div>
-              )}
-            </ul>
+            <VideoControls
+              activeCallViewTab={activeCallViewTab}
+              call={call}
+              isMuted={isMuted}
+              isParticipantsListOpen={isParticipantsListOpen}
+              toggleParticipantsList={toggleParticipantsList}
+              canShareScreen={canShareScreen}
+              conversation={conversation}
+              mediaDevicesHandler={mediaDevicesHandler}
+              minimize={minimize}
+              leave={leave}
+              toggleMute={toggleMute}
+              toggleScreenshare={toggleScreenshare}
+              toggleCamera={toggleCamera}
+              toggleIsHandRaised={toggleIsHandRaised}
+              switchMicrophoneInput={switchMicrophoneInput}
+              switchSpeakerOutput={switchSpeakerOutput}
+              switchBlurredBackground={switchBlurredBackground}
+              switchCameraInput={switchCameraInput}
+              setActiveCallViewTab={setActiveCallViewTab}
+              setMaximizedParticipant={setMaximizedParticipant}
+              sendEmoji={sendEmoji}
+            />
           </div>
         )}
       </div>
